Fixed race condition in Contract.tx_call @ aeternity/contract.py by making the method block until the call transaction is mined
This also applies in non-blocking mode.

Added a synchronous flag to the method, when this is false it will not
wait for the tx to be mined, but also return None in place of the
result of the contract call.

Can't have it both ways... function must either wait or not return the
call result. Previously in nonblocking mode it would throw an exception
(usually)